requirements.txt
pandas==1.3.3
numpy==1.21.2
torch==1.13.1
transformers==4.30.2
requests==2.26.0
scikit-image==0.18.3
Pillow==8.4.0
findspark==1.4.2
IPython==7.20.0
matplotlib==3.3.4